Separate create and display person in a new Class
